+++ CHANGES.rst
@@ -8,6 +8,11 @@ In development
   be passed directly to ``Memory.reduce_size``.
   https://github.com/joblib/joblib/pull/1569
 
+- Extend functionality of the ``check_call_in_cache`` method to now also
+  check against cache validity. Before, it would only check for a given call
+  if it is in cache memory.
+  https://github.com/joblib/joblib/pull/1584
+
 Release 1.4.2 -- 2024/05/02
 ---------------------------