+++ CHANGES.rst
@@ -8,6 +8,9 @@ In development
   be passed directly to ``Memory.reduce_size``.
   https://github.com/joblib/joblib/pull/1569
 
+- Support for Python 3.13 free-threaded has been added.
+  https://github.com/joblib/joblib/pull/1589
+
 Release 1.4.2 -- 2024/05/02
 ---------------------------